Rework Translations
Made translations modular instead of a single file
