--- frontend/languages.json
@@ -1,73 +1,0 @@
-{
-	"en": {
-		"home": "Home",
-		"play": "Play",
-		"profile": "Profile",
-		"login": "Login",
-		"signUp": "Sign Up",
-		"friends": "Friends",
-		"friendsSidebar": "Friends Menu",
-		"addFriend": "Add Friend",
-		"profileTitle": "Your Profile",
-		"statsTitle": "Stats",
-		"gamesPlayed": "Games Played:",
-		"wins": "Wins:",
-		"losses": "Losses:",
-		"responsive-title": "Pong Master",
-		"responsive-subtitle": "Step up to the table and unleash your inner champion with Pongmasters",
-		"tournament": "Tournament",
-		"friendsButton": "Friends",
-		"profileButton": "Profile",
-		"friendsTab": "Friends",
-		"requestsTab": "Requests",
-		"addFriendsTab": "Add Friends"
-	},
-	"es": {
-		"home": "Inicio",
-		"play": "Jugar",
-		"tournaments": "Torneos",
-		"profile": "Perfil",
-		"login": "Iniciar sesión",
-		"signUp": "Registrarse",
-		"friends": "Amigos",
-		"friendsSidebar": "Menú de Amigos",
-		"addFriend": "Agregar amigo",
-		"profileTitle": "Tu Perfil",
-		"statsTitle": "Estadísticas",
-		"gamesPlayed": "Juegos Jugados: 50",
-		"wins": "Victorias: 30",
-		"losses": "Derrotas: 20",
-		"responsive-title": "Pong Master",
-		"responsive-subtitle": "Pasa al siguiente nivel y desata tu campeón interior con Pongmasters",
-		"tournament": "Torneo",
-		"friendsButton": "Amigos",
-		"profileButton": "Perfil",
-		"friendsTab": "Amigos",
-		"requestsTab": "Solicitudes",
-		"addFriendsTab": "Agregar amigos"
-	},
-	"de": {
-		"home": "Startseite",
-		"play": "Spielen",
-		"tournaments": "Turniere",
-		"profile": "Profil",
-		"login": "Anmelden",
-		"signUp": "Registrieren",
-		"friends": "Freunde",
-		"friendsSidebar": "Freunde Menü",
-		"addFriend": "Freund hinzufügen",
-		"profileTitle": "Dein Profil",
-		"statsTitle": "Statistiken",
-		"gamesPlayed": "Gespielte Spiele: 50",
-		"wins": "Siege: 30",
-		"losses": "Niederlagen: 20",
-		"responsive-title": "Pong Master",
-		"responsive-subtitle": "Tritt an den Tisch und entfessle deinen inneren Champion mit Pongmasters",
-		"tournament": "Turnier",
-		"friendsButton": "Freunde",
-		"profileButton": "Profil",
-		"friendsTab": "Freunde",
-		"requestsTab": "Anfragen",
-		"addFriendsTab": "Freunde hinzufügen"
-	}
-}
