+++ frontend/pages/home.html
@@ -3,17 +3,17 @@
 	<div class="col h-50 d-flex bg-home justify-content-center align-items-center">
 		<div class="text-center">
 			<div class="row mb-4">
-				<h2 class="responsive-title" data-translate="responsive-title">Pong Master</h2>
+				<h2 class="responsive-title" data-translate="pongMaster">Pong Master</h2>
 			</div>
 			<div class="row">
-				<button onclick="loadPage('play')" class="btn btn-primary mb-3" data-translate="playButton">Play</button>
+				<button onclick="loadPage('play')" class="btn btn-primary mb-3" data-translate="playNow">Play now</button>
 			</div>
 		</div>
 	</div>
 	<div class="container-fluid text-center">
 		<div class="row mb-3 d-flex justify-content-center align-items-center">
 			<div class="text-center">
-				<h3 class="responsive-subtitle" data-translate="responsive-subtitle">Step up to the table and unleash your inner champion with Pongmasters</h3>
+				<h3 class="responsive-subtitle" data-translate="welcomeMessage">Step up to the table and unleash your inner champion with Pong Master</h3>
 			</div>
 		</div>
 	</div>
