+++ frontend/translations/home/de.json
@@ -1,0 +1,4 @@
+{
+	"playNow":				"Jetzt spielen",
+	"welcomeMessage":	"Tritt an den Tisch und entfessle deinen inneren Champion mit Pong Meister"
+}
