+++ frontend/translations/home/en.json
@@ -1,0 +1,4 @@
+{
+	"playNow":				"Play Now",
+	"welcomeMessage":	"Step up to the table to unleash your inner champion and be the Pong Master"
+}
