+++ frontend/translations/home/es.json
@@ -1,0 +1,4 @@
+{
+	"playNow":				"Jugar ahora",
+	"welcomeMessage":	"Preparate para desatar tu campeón interior, y vuelvete el rey del Pong"
+}
